doc:新增HTTPFuzzerPage.ts用来存放该页面的默认值,更换默认值的导出文件
