+++ app/renderer/src/main/src/pages/fuzzer/HTTPFuzzerHotPatch.tsx
@@ -1,16 +1,13 @@
 import React, {useEffect, useState} from "react"
-import {Button, Checkbox, Form, Popconfirm, Popover, Space, Tag, Tooltip} from "antd"
+import {Form, Space} from "antd"
 import {YakEditor} from "../../utils/editors"
 import {callCopyToClipboard} from "../../utils/basic"
-import {showDrawer, showModal} from "../../utils/showModal"
 import {AutoCard} from "../../components/AutoCard"
 import {getRemoteValue, setRemoteValue} from "@/utils/kv"
 import {useGetState, useMemoizedFn} from "ahooks"
 import {RefreshIcon} from "@/assets/newIcon"
 import {ExclamationCircleOutlined, FullscreenOutlined} from "@ant-design/icons/lib"
-import {SelectOne} from "@/utils/inputUtil"
 import {YakitButton} from "@/components/yakitUI/YakitButton/YakitButton"
-import {YakitCheckbox} from "@/components/yakitUI/YakitCheckbox/YakitCheckbox"
 import {YakitRadioButtons} from "@/components/yakitUI/YakitRadioButtons/YakitRadioButtons"
 import {YakitPopover} from "@/components/yakitUI/YakitPopover/YakitPopover"
 import {YakitPopconfirm} from "@/components/yakitUI/YakitPopconfirm/YakitPopconfirm"
@@ -18,8 +15,8 @@ import styles from "./HTTPFuzzerHotPatch.module.scss"
 import {showYakitDrawer} from "@/components/yakitUI/YakitDrawer/YakitDrawer"
 import {yakitNotify} from "@/utils/notification"
 import {OutlineXIcon} from "@/assets/icon/outline"
-import {WEB_FUZZ_HOTPATCH_CODE} from "./HTTPFuzzerPage"
 import {YakitModalConfirm} from "@/components/yakitUI/YakitModal/YakitModalConfirm"
+import {WEB_FUZZ_HOTPATCH_CODE} from "@/defaultConstants/HTTPFuzzerPage"
 
 export interface HTTPFuzzerHotPatchProp {
     onInsert?: (s: string) => any
