+++ app/renderer/src/main/src/pages/fuzzer/components/ShareImportExportData/index.tsx
@@ -6,7 +6,7 @@ import {showYakitModal} from "@/components/yakitUI/YakitModal/YakitModalConfirm"
 import {onImportShare} from "@/pages/fuzzer/components/ShareImport"
 import styles from "./index.module.scss"
 import {useStore} from "@/store"
-import {success, yakitNotify, yakitFailed, yakitInfo, warn} from "@/utils/notification"
+import {success, yakitNotify, yakitFailed, warn} from "@/utils/notification"
 import CopyToClipboard from "react-copy-to-clipboard"
 import {NetWorkApi} from "@/services/fetch"
 import {API} from "@/services/swagger/resposeType"
@@ -21,12 +21,13 @@ import {saveABSFileAnotherOpen} from "@/utils/openWebsite"
 import {Uint8ArrayToString, StringToUint8Array} from "@/utils/str"
 import {NewHTTPPacketEditor} from "@/utils/editors"
 import {YakitRoute} from "@/routes/newRoute"
-import {DefFuzzerTableMaxData, FuzzerRequestProps} from "../../HTTPFuzzerPage"
+import {FuzzerRequestProps} from "../../HTTPFuzzerPage"
 import {AdvancedConfigValueProps} from "../../HttpQueryAdvancedConfig/HttpQueryAdvancedConfigType"
 import emiter from "@/utils/eventBus/eventBus"
 import {randomString} from "@/utils/randomUtil"
 import {generateGroupId} from "@/pages/layout/mainOperatorContent/MainOperatorContent"
 import {MultipleNodeInfo} from "@/pages/layout/mainOperatorContent/MainOperatorContentType"
+import {DefFuzzerTableMaxData} from "@/defaultConstants/HTTPFuzzerPage"
 
 const {ipcRenderer} = window.require("electron")
