+++ app/renderer/src/main/src/store/pageInfo.ts
@@ -2,13 +2,14 @@ import {AdvancedConfigValueProps} from "@/pages/fuzzer/HttpQueryAdvancedConfig/H
 import {YakitRoute} from "@/routes/newRoute"
 import {subscribeWithSelector, persist, StorageValue} from "zustand/middleware"
 import debounce from "lodash/debounce"
-import {AdvancedConfigShowProps, defaultAdvancedConfigValue, defaultPostTemplate} from "@/pages/fuzzer/HTTPFuzzerPage"
+import {AdvancedConfigShowProps} from "@/pages/fuzzer/HTTPFuzzerPage"
 import {yakitNotify} from "@/utils/notification"
 import {RemoteGV} from "@/yakitGV"
 import {setRemoteProjectValue} from "@/utils/kv"
 import cloneDeep from "lodash/cloneDeep"
 import {createWithEqualityFn} from "zustand/traditional"
 import {HybridScanControlAfterRequest, HybridScanModeType} from "@/models/HybridScan"
+import {defaultAdvancedConfigValue, defaultPostTemplate} from "@/defaultConstants/HTTPFuzzerPage"
 
 /**
  * @description 页面暂存数据
@@ -83,13 +84,6 @@ export interface PluginBatchExecutorPageInfoProps {
     request: Uint8Array
     /**执行任务的状态 */
     hybridScanMode: HybridScanModeType
-}
-export const defaultWebFuzzerPageInfo: WebFuzzerPageInfoProps = {
-    pageId: "",
-    advancedConfigValue: cloneDeep(defaultAdvancedConfigValue),
-    advancedConfigShow: null,
-    request: defaultPostTemplate,
-    variableActiveKeys: undefined
 }
 
 export interface WebFuzzerPageInfoProps {
